+++ packages/vscode-extension/src/shared/constants.ts
@@ -1,9 +1,0 @@
-export {
-  GESTURE_MAP,
-  GESTURE_KEYS,
-  EASING_MAP,
-  TRANSITION_KEYWORDS,
-  VIEWPORT_KEYWORDS,
-  DRAG_KEYWORDS,
-  LAYOUT_KEYWORDS,
-} from "../../../motionwind/src/constants.js";